Show SrcLoc for clicked piece in Debug Mode
When a user clicks on a picture consisting of one piece (polygon,
path,sector, etc) the location that piece was created in code will
show on message log in bottom right.

Since the method to determine which piece was clicked will change,
this part of the code is not fully implemented. The eventual goal is
for debugMode functionality to be built into the UI so users do not
have to explicitly run programs as debugMode pic.